+++ docs/CHANGELOG.md
@@ -5,6 +5,7 @@ Changelog
 --------------------
 - Fix #51: Initialize module content class
 - Enh #52: Tests for `next` version
+- Fix #53: Fix space searching with updated default permission
 
 1.0.3 (June 2, 2023)
 --------------------